Polished off the physics in bullet.c and start playerControl.h

+++ lab9_project/bullet.h
@@ -9,10 +9,10 @@ typedef struct {
   double x_vel;
   double y_vel;
 
-  uint16_t x_current;
-  uint16_t y_current;
+  double x_current;
+  double y_current;
 
-  int16_t wind;
+  double wind;
 
   bool dead;
 
@@ -24,7 +24,7 @@ typedef struct {
 void bullet_init_dead(bullet_t *bullet);
 
 // Initialize the bullet state machine
-void bullet_init(bullet_t *bullet, uint16_t x_origin, uint16_t y_origin, uint16_t power, double angle, int16_t wind);
+void bullet_init(bullet_t *bullet, uint16_t x_origin, uint16_t y_origin, double power, double angle, double wind);
 
 
 ////////// State Machine TICK Function //////////
